Refactor Firebase authentication imports and paths
