+++ src/services/firebase/firebase-config.ts
@@ -1,6 +1,7 @@
 // Import the functions you need from the SDKs you need
 import { getAnalytics } from "firebase/analytics"
 import { initializeApp } from "firebase/app"
+import { getAuth } from "firebase/auth"
 import { getFirestore } from "firebase/firestore"
 // TODO: Add SDKs for Firebase products that you want to use
 // https://firebase.google.com/docs/web/setup#available-libraries
@@ -22,6 +23,9 @@ export const app = initializeApp(firebaseConfig)
 const analytics = getAnalytics(app)
 
 export const db = getFirestore(app)
+
+// Create a Firebase authentication reference
+export const auth = getAuth(app)
 
 // firebase login
 // firebase init
